Simplify build matrix
Remove the default `experimental` entry in the build matrix, and instead
mark Ruby versions and/or gemfiles that are experimental as `true` in
the `include` section.

This works really well, since the default value for `experimental` is
`false`, which is basically the same as `undefined`, so it doesn't need
to be declared explicitly. I made sure that the `continue-on-error` is
properly assigned a boolean value by using the `!!` operator, to be safe.

As for the `includes` section, the documentation states that:

> the key:value pairs in the object will be added to each of the matrix
combinations if none of the key:value pairs overwrite any of the
original matrix values.

This means that if we add `experimental: true` for a specific matrix
item, then that will be added to all the matrix entries that match that
matrix item.

Using this, we can mark all matrix entries that use `Gemfile-rails-main`
gemfile as being experimental, without having to enumerate all the Ruby
versions again.