Protobuf DSL: generate namespaced enum constants on root
The Protobuf DSL compiler was causing errors by generating un-namespaced
constants for enums. In certain cases, this could result in a constant
conflicting with a reserved Ruby word, such as "BEGIN." Instead,
we generate the constants with the correct namespace outside of the
class or module, ensuring that RBIs never contain code such as
"BEGIN = 3" which is invalid Ruby syntax.

Co-authored-by: Alexandre Terrasa <alexandre.terrasa@shopify.com>