Stop returning names of `TypeVariableModule` instances
Sorbet runtime should never need to look up the name of a
`TypeVariableModule` instance, since that type should always be coerced
to a `Tapioca::TypeVariable` instance by the point the `name` method is
called.

The fact that we were doing this was redundant and was preventing us
from detecting instances where a `TypeVariableModule` instance was not
being properly coerced to a `Tapioca::TypeVariable` instance.