Avoid swallowing errors in (most) CLI tests
By default `exe/tapioca` will ignore many errors, including typechecking
ones. This is because we want to generate as much RBI as possible, even
if there are errors in the codebase.

We do want to surface more errors in the CLI tests, however, so this PR
is making CLI tests throw by default, while providing an option to
ignore it explicitly in cases where the updates to the test setup would
not be worth the hassle.

See https://github.com/Shopify/tapioca/pull/1277#issuecomment-1326769461