Do not try to parse the return of `say_error`
This error was found thanks to #1287 and is actually quite fun.

The `trees` variable is untyped since we use `executor.run_in_parallel`
and at runtime we end up returning an array of `T.any(RBI::Tree, T::Private::Types::Void::VOID)`
when we encounter a parse error during RBI parsing.

Because it's untyped, no static error comes up with `srb tc` but at runtime
we actually try to pass a `T::Private::Types::Void::VOID` to `visit_all`.

We need to skip this case and return `nil` instead so the `visit_all` will ignore it.

Signed-off-by: Alexandre Terrasa <alexandre.terrasa@shopify.com>