Improve eager loading Rails engines with Zeitwerk
RBI generation for Rails engines is currently implemented with a best effort approach. We identify every file in the application's eager_load_path and attempt to require those files. This does not account for dependency order, so we keep track of any files that fail in the process and then try to require them again once all the other files are loaded.

This approach is not exhaustive and can sometimes miss constants defined in engines. For example, if an engine defines a constant (e.g. Turbo::Streams) as a namespace for subconstants, but never on its own, the current approach will never generate type annotations for that constant.

This implementation improves engine eager loading when Zeitwerk is present in the Rails application. Rather than relying on the best effort approach, we use the behavior already implemented in Zeitwerk to eager load every file in the application's eager_load_path, which is much more reliable. The resulting RBI will contain constants defined in the engine that wouldn't have been found by the previous approach.

If Zeitwerk is not present, then we fall back to the best effort approach that we were already using.