Use Ruby 3.2 `Class#attached_object` method
Partially thanks to our work on Tapioca, Ruby 3.2 introduced the
`Class#attached_object` method, which fetches the attached object of a
singleton class. We are currently doing this in tapioca by examining all of
ObjectSpace, which is not very elegant or performant.

This changes the implementation of `attached_class_of` in
`Runtime::Reflection` to just call the `attached_object` method if it's
defined on the class being passed in. If this method is not defined
(because we're running Ruby 3.1 or older), then we fall back to the previous
implementation.