Add global callbacks to AASM compiler
AASM state machines allow registering global callbacks like
after_all_transitions, before_all_transactions, and so on. They accept a block
that runs in the context of the record. The gem compiler creates these
callbacks, but the block is of course bound to T.untyped, requiring a cast in
the application code. This change extends the AASM compiler to also generate
the global callbacks (in addition to the event callbacks it already generates).
Since the resulting methods are typed with blocks bound to the record class,
the correct type appears in application code.