Only enable `require: false` override inside a block
This commit limits the `require: false` override to only be enabled
during the execution of a block passed to the `.override_require_false`
method on `AutoRequireHook` module. This is only done while we are doing
essentially what is a `Bundler.require` call.

In all other cases, the `AutoRequireHook` module will be prepended to `Bundler::Dependency` but since it is not enabled, it will not do anything inside the `autorequire` method.