Simplify protobuf compiler default field values
The no point in trying to construct elaborate default values for
the initializer fields, since in the RBI file, the default value is meaningless. What matters is the type of the field and if the field is optional or not. Thus, we can easily simplify the code and generate default values as `T.unsafe(nil)`.