Skip compiling constants with alias in namespace
When compiling a gem, it is possible for the gem to define an alias that
includes an alias from another gem in its namespace. For example:

In gem foo:
```
module Foo; end
F = Foo
```

In gem bar:
```
module Foo
  module Bar; end
end

F::B = Foo::Bar
```

Because the alias `F` is not originally defined in the gem `bar`,
tapioca will not skip writing that alias to the RBI file of `bar`, which
will cause a Sorbet error because the alias `F` will be defined in the
RBI files for both `foo` and `bar`.

Now, tapioca will not generate RBI for aliases originally defined in
other gems.