Change how extensions register themselves
Extensions cannot guard against the constants being available at load time, since we load extensions before we load the application. Instead, we need to register load hooks that will be called when the relevant constants are available.

For now, this only works for Active Record in every case, since Active Record has a lazy load hook that we can use. For Frozen Record and Kredis we need to rely on the Rails load hooks to delayed trigger of the extension loading. But, that won't be available outside of Rails.