Automatic stdout/stderr assertion updates
When I was working on the ripper compiler, I found it difficult to
have to update every single assertion that is assertion against
stdout/stderr of tapioca. Because it's not assertion against a
subset, any change to the logic will cause the entire test to fail
and have to be updated.

These are more like snapshot tests than unit tests. Therefore I
built a way to update all of the snapshots in one go. It's not
meant to be comprehensive or even tested, but I found it very
useful as opposed to trying to update every fixture myself, so I
figured I would upstream it in case there was interest.