Add overloads to active record fixture methods
This is to discriminate between single argument calls which return a
single model, e.g. User and multi argument calls which return an array
of the model, e.g. T::Array[User].

This is the precursor to actually typing the return values since
otherwise we would not be able to type it correctly.