Parse DSL compiler options into YAML if they are supplied as strings
Instead of erroring out when the compiler options are not supplied as hashes, we could try to first parse them as YAML (if they are supplied as strings) and only then error if any of the compiler options isn't a hash.

This allows users to supply compiler options as strings in the `tapioca` CLI as:
```
dsl Post --compiler-options='ActiveRecordColumns:{types: untyped}'
```
which is more user-friendly than not being able to pass these options on the command line.